+++ src/data/appData.ts
@@ -463,7 +463,7 @@ const DEFAULT_DEFICITS: AppDeficit[] = [
     "kontext": "io",
     "isPflicht": true,
     "isBooster": false,
-    "normRefs": ["VSS SN 640 075", "VSS 41 723"],
+    "normRefs": ["VSS SN 640 075", "SN 641 723"],
     "nameI18n": {
       "de": "Fehlende Absenkung",
       "fr": "Abaissement absent",
@@ -494,7 +494,7 @@ const DEFAULT_DEFICITS: AppDeficit[] = [
     "kontext": "io",
     "isPflicht": true,
     "isBooster": false,
-    "normRefs": ["SN 640 273", "VSS 41 723"],
+    "normRefs": ["SN 640 273", "SN 641 723"],
     "nameI18n": {
       "de": "Sichtbehinderung Hecke",
       "fr": "Obstruction par haie",
@@ -532,7 +532,7 @@ const DEFAULT_DEFICITS: AppDeficit[] = [
     "kontext": "ao",
     "isPflicht": true,
     "isBooster": false,
-    "normRefs": ["SN 640 238", "VSS 41 723"],
+    "normRefs": ["SN 640 238", "SN 641 723"],
     "nameI18n": {
       "de": "Unterbrochener Radstreifen",
       "fr": "Piste cyclable interrompue",
@@ -558,7 +558,7 @@ const DEFAULT_DEFICITS: AppDeficit[] = [
     "kontext": "io",
     "isPflicht": false,
     "isBooster": true,
-    "normRefs": ["SN 640 852", "VSS 41 723"],
+    "normRefs": ["SN 640 852", "SN 641 723"],
     "nameI18n": {
       "de": "Fehlende Wartelinie",
       "fr": "Ligne d'attente manquante",
@@ -590,7 +590,7 @@ const DEFAULT_DEFICITS: AppDeficit[] = [
     },
     "isPflicht": true,
     "isBooster": true,
-    "normRefs": ["VSS 41 723"],
+    "normRefs": ["SN 641 723"],
     "verortung": {
       "typ": "punkt",
       "position": { "theta": 271.99698118750723, "phi": 132.0718240138629 },
@@ -611,7 +611,7 @@ const DEFAULT_DEFICITS: AppDeficit[] = [
     },
     "isPflicht": true,
     "isBooster": false,
-    "normRefs": ["VSS 41 723"],
+    "normRefs": ["SN 641 723"],
     "verortung": {
       "typ": "polygon",
       "punkte": [